# erp-domain/openapi.yaml
openapi: 3.0.1
info:
  title: BigBright ERP Domain API
  description: Core APIs for multi-company ERP platform
  license:
    name: Proprietary
  version: v1
...
servers:
- url: http://localhost:51904
  description: Generated server url
paths:
...
  /api/v1/admin/users/{id}:
    put:
      tags:
      - admin-user-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserDto"
    delete:
      tags:
      - admin-user-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/admin/settings:
    get:
      tags:
      - admin-settings-controller
      operationId: getSettings
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSystemSettingsDto"
    put:
      tags:
      - admin-settings-controller
      operationId: updateSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemSettingsUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSystemSettingsDto"
...
  /api/v1/accounting/catalog/products/{id}:
    put:
      tags:
      - accounting-catalog-controller
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProductionProductDto"
...
  /api/v1/auth/mfa/disable:
    post:
      tags:
      - mfa-controller
      operationId: disable
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MfaDisableRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMfaStatusResponse"
...
  /api/v1/admin/users:
    get:
      tags:
      - admin-user-controller
      operationId: list_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListUserDto"
    post:
      tags:
      - admin-user-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseUserDto"
...
  /api/v1/admin/notify:
    post:
      tags:
      - admin-settings-controller
      operationId: notifyUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminNotifyRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
...
  /api/v1/accounting/catalog/products:
    get:
      tags:
      - accounting-catalog-controller
      operationId: listProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListProductionProductDto"
    post:
      tags:
      - accounting-catalog-controller
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseProductionProductDto"
  /api/v1/accounting/catalog/products/bulk-variants:
    post:
      tags:
      - accounting-catalog-controller
      operationId: createVariants
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkVariantRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseBulkVariantResponse"
  /api/v1/accounting/catalog/import:
    post:
      tags:
      - accounting-catalog-controller
      operationId: importCatalog
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCatalogImportResponse"
...
  /api/v1/admin/users/{id}/unsuspend:
    patch:
      tags:
      - admin-user-controller
      operationId: unsuspend
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/admin/users/{id}/suspend:
    patch:
      tags:
      - admin-user-controller
      operationId: suspend
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/admin/users/{id}/mfa/disable:
    patch:
      tags:
      - admin-user-controller
      operationId: disableMfa
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    UpdateUserRequest:
      required:
      - displayName
      type: object
      properties:
        displayName:
          type: string
        companyIds:
          type: array
          items:
            type: integer
            format: int64
        roles:
          type: array
          items:
            type: string
        enabled:
          type: boolean
    ApiResponseUserDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/UserDto"
        timestamp:
          type: string
          format: date-time
    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        email:
          type: string
        displayName:
          type: string
        enabled:
          type: boolean
        mfaEnabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        companies:
          type: array
          items:
            type: string
    SystemSettingsUpdateRequest:
      type: object
      properties:
        allowedOrigins:
          type: array
          items:
            type: string
        autoApprovalEnabled:
          type: boolean
        periodLockEnforced:
          type: boolean
        mailEnabled:
          type: boolean
        mailFromAddress:
          type: string
        mailBaseUrl:
          type: string
        sendCredentials:
          type: boolean
        sendPasswordReset:
          type: boolean
    ApiResponseSystemSettingsDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/SystemSettingsDto"
        timestamp:
          type: string
          format: date-time
    SystemSettingsDto:
      type: object
      properties:
        allowedOrigins:
          type: array
          items:
            type: string
        autoApprovalEnabled:
          type: boolean
        periodLockEnforced:
          type: boolean
        mailEnabled:
          type: boolean
        mailFromAddress:
          type: string
        mailBaseUrl:
          type: string
        sendCredentials:
          type: boolean
        sendPasswordReset:
          type: boolean
...
    ProductUpdateRequest:
      type: object
      properties:
        productName:
          type: string
        category:
          type: string
        defaultColour:
          type: string
        sizeLabel:
          type: string
        unitOfMeasure:
          type: string
        basePrice:
          type: number
        gstRate:
          type: number
        minDiscountPercent:
          type: number
        minSellingPrice:
          type: number
        metadata:
          type: object
          additionalProperties:
            type: object
    ApiResponseProductionProductDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/ProductionProductDto"
        timestamp:
          type: string
          format: date-time
    ProductionProductDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicId:
          type: string
          format: uuid
        brandId:
          type: integer
          format: int64
        brandName:
          type: string
        brandCode:
          type: string
        productName:
          type: string
        category:
          type: string
        defaultColour:
          type: string
        sizeLabel:
          type: string
        unitOfMeasure:
          type: string
        skuCode:
          type: string
        active:
          type: boolean
        basePrice:
          type: number
        gstRate:
          type: number
        minDiscountPercent:
          type: number
        minSellingPrice:
          type: number
        metadata:
          type: object
          additionalProperties:
            type: object
...
    ApiResponseString:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: string
        timestamp:
          type: string
          format: date-time
...
    MfaDisableRequest:
      type: object
      properties:
        code:
          type: string
        recoveryCode:
          type: string
    ApiResponseMfaStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/MfaStatusResponse"
        timestamp:
          type: string
          format: date-time
    MfaStatusResponse:
      type: object
      properties:
        enabled:
          type: boolean
...
    CreateUserRequest:
      required:
      - companyIds
      - displayName
      - email
      - roles
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        displayName:
          type: string
        companyIds:
          type: array
          items:
            type: integer
            format: int64
        roles:
          type: array
          items:
            type: string
...
    AdminNotifyRequest:
      required:
      - body
      - subject
      - to
      type: object
      properties:
        to:
          type: string
        subject:
          type: string
        body:
          type: string
...
    ProductCreateRequest:
      required:
      - category
      - productName
      type: object
      properties:
        brandId:
          type: integer
          format: int64
        brandName:
          type: string
        brandCode:
          type: string
        productName:
          type: string
        category:
          type: string
        defaultColour:
          type: string
        sizeLabel:
          type: string
        unitOfMeasure:
          type: string
        customSkuCode:
          type: string
          writeOnly: true
        basePrice:
          type: number
        gstRate:
          type: number
        minDiscountPercent:
          type: number
        minSellingPrice:
          type: number
        metadata:
          type: object
          additionalProperties:
            type: object
    BulkVariantRequest:
      required:
      - baseProductName
      - category
      - colors
      - sizes
      type: object
      properties:
        brandId:
          type: integer
          format: int64
        brandName:
          type: string
        brandCode:
          type: string
        baseProductName:
          type: string
        category:
          type: string
        colors:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        unitOfMeasure:
          type: string
        skuPrefix:
          type: string
        basePrice:
          type: number
        gstRate:
          type: number
        minDiscountPercent:
          type: number
        minSellingPrice:
          type: number
        metadata:
          type: object
          additionalProperties:
            type: object
    ApiResponseBulkVariantResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/BulkVariantResponse"
        timestamp:
          type: string
          format: date-time
    BulkVariantResponse:
      type: object
      properties:
        created:
          type: integer
          format: int32
        skippedExisting:
          type: integer
          format: int32
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductionProductDto"
    ApiResponseCatalogImportResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/CatalogImportResponse"
        timestamp:
          type: string
          format: date-time
    CatalogImportResponse:
      type: object
      properties:
        rowsProcessed:
          type: integer
          format: int32
        brandsCreated:
          type: integer
          format: int32
        productsCreated:
          type: integer
          format: int32
        productsUpdated:
          type: integer
          format: int32
        rawMaterialsSeeded:
          type: integer
          format: int32
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportError"
    ImportError:
      type: object
      properties:
        rowNumber:
          type: integer
          format: int64
        message:
          type: string
...
    ApiResponseListProductionProductDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProductionProductDto"
        timestamp:
          type: string
          format: date-time
...
    ApiResponseListUserDto:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserDto"
        timestamp:
          type: string
          format: date-time
...