# erp-domain/src/main/java/com/bigbrightpaints/erp/modules/admin/dto/SystemSettingsDto.java
package com.bigbrightpaints.erp.modules.admin.dto;

import java.util.List;

public record SystemSettingsDto(
        List<String> allowedOrigins,
        boolean autoApprovalEnabled,
        boolean periodLockEnforced,
        boolean mailEnabled,
        String mailFromAddress,
        String mailBaseUrl,
        boolean sendCredentials,
        boolean sendPasswordReset
) {}